refactor: fix pointer truncation, modifier keys, CPU resize, and deduplicate event dispatch
- Fix lossy pointer-to-u32 cast to u64 for 64-bit safety in surface ID
- Map right-hand modifier keys to RShift/RControl/RAlt/RWin instead of left variants
- Implement CPU buffer resize by recreating CpuFallbackState with new dimensions
- Extract handle_process_event_result() helper to deduplicate 3 identical dispatch blocks

Autoreview: dll/src/desktop/shell2/linux/wayland/mod.rs

+++ dll/src/desktop/shell2/linux/gnome_menu/manager.rs
@@ -254,7 +254,7 @@ impl GnomeMenuManager {
     /// This function exists for API consistency but doesn't set any properties.
     pub fn set_window_properties_wayland(
         &self,
-        _surface_id: u32,
+        _surface_id: u64,
         app_id: &Option<String>,
     ) -> Result<(), GnomeMenuError> {
         debug_log("Wayland menu integration (using app_id matching)");
